add babel transform object assign
