+++ .babelrc
@@ -3,6 +3,7 @@
   "plugins": [
     "transform-es2015-modules-commonjs",
     "array-includes",
+    "transform-object-assign",
     "transform-object-rest-spread"
 ]
 }
